requirements.txt
# AutriFix backend — Django 5 + DRF + async stack
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
whitenoise==6.9.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.10

# Location (lat/lng + geodesic distance; maps on the client)
geopy==2.4.1

# Real-time
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Async tasks & cache
celery==5.4.0
redis==5.2.1
django-redis==5.4.0

# Media
django-cloudinary-storage==0.3.0
cloudinary==1.41.0

# Utils
Pillow==11.1.0
python-json-logger==3.2.1

# Dev
gunicorn==23.0.0
